requirements.txt
amqp==2.2.2
beautifulsoup4==4.6.0
billiard==3.5.0.3
blinker==1.4
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
flake8==3.5.0
flake8-docstrings==1.1.0
flake8-polyfill==1.0.1
Flask==0.12.2
Flask-Mail==0.9.1
Flask-Uploads==0.2.1
hacking==1.0.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
kombu==4.1.0
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.13.3
pandas==0.22.0
pbr==3.1.1
pep8==1.5.7
pika==0.11.2
pycodestyle==2.3.1
pydocstyle==2.1.1
pyflakes==1.6.0
pymongo==3.6.0
python-dateutil==2.6.1
pytz==2017.3
redis==2.10.6
requests==2.18.4
six==1.11.0
snowballstemmer==1.2.1
urllib3==1.22
uWSGI==2.0.15
vine==1.1.4
virtualenv==15.1.0
Werkzeug==0.14.1
